+++ src/cli/dcos_aws/commands/create.py
@@ -163,6 +163,7 @@ def create(
         Distribution.CENTOS_7: 'centos',
         Distribution.COREOS: 'core',
         Distribution.UBUNTU_16_04: 'ubuntu',
+        Distribution.RHEL_7: 'ec2-user',
     }
 
     distribution = LINUX_DISTRIBUTIONS[linux_distribution]