+++ src/dcos_e2e/distributions.py
@@ -15,3 +15,4 @@ class Distribution(Enum):
     CENTOS_7 = 1
     COREOS = 2
     UBUNTU_16_04 = 3
+    RHEL_7 = 4